# docs/apidoc.json
{"openapi":"3.0.0","info":{"title":"API document","version":"v1"},"servers":[{"url":"http://localhost:5000/api/v1"}],"tags":[{"name":"Test","description":"Test operators"},{"name":"Auth","description":"Auth operators"}],"paths":{"/tests/:id":{"delete":{"tags":["Test"],"description":"action desccription","operationId":"update","summary":"update","security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","description":"paramester description","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"name":{"minLength":1,"maxLength":255,"type":"string"},"description":{"minLength":1,"maxLength":255,"type":"string"}},"type":"object","required":["name"]}}}}},"actionName":"deleteOne","get":{"tags":["Test"],"description":"action desccription","operationId":"getOne","summary":"getOne","security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","description":"paramester description","required":false,"schema":{"type":"string"}}]},"patch":{"tags":["Test"],"description":"action desccription","operationId":"deleteOne","summary":"deleteOne","security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","description":"paramester description","required":false,"schema":{"type":"string"}}]}},"/tests/":{"get":{"tags":["Test"],"description":"action desccription","operationId":"getList","summary":"getList","security":[{"BearerAuth":[]}],"parameters":[{"in":"query","description":"paramester description","required":false,"content":{"application/json":{"schema":{"properties":{"filter":{"properties":{"type":{"enum":["AND","OR"],"type":"string"},"filters":{"items":{"properties":{"property":{"type":"string"},"operator":{"type":"string","enum":["eq","lt","lte","gt","gte","ne","in","like","between"]},"value":{}},"type":"object","required":["property","operator"]},"type":"array"}},"type":"object","required":["type","filters"]},"sorts":{"items":{"type":"string"},"type":"array"},"limit":{"minimum":1,"type":"number"},"page":{"minimum":1,"type":"number"}},"type":"object"}}}}]},"patch":{"tags":["Test"],"description":"action desccription","operationId":"deleteList","summary":"deleteList","security":[{"BearerAuth":[]}],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"properties":{"ids":{"items":{"type":"string"},"type":"array"}},"type":"object","required":["ids"]}}}}},"post":{"tags":["Test"],"description":"action desccription","operationId":"create","summary":"create","security":[{"BearerAuth":[]}],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"properties":{"name":{"minLength":1,"maxLength":255,"type":"string"},"description":{"minLength":1,"maxLength":255,"type":"string"}},"type":"object","required":["name"]}}}}}},"/auth/facebook/callback":{"actionName":"facebookCallback","get":{"tags":["Auth"],"description":"action desccription","operationId":"facebookCallback","summary":"facebookCallback","security":[{"BearerAuth":[]}],"parameters":[]}},"/auth/register":{"post":{"tags":["Auth"],"description":"action desccription","operationId":"register","summary":"register","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"properties":{"email":{"minLength":1,"maxLength":255,"type":"string","format":"email"},"password":{"minLength":6,"maxLength":20,"type":"string"},"roleId":{"format":"uuid","type":"string"}},"type":"object","required":["email","password","roleId"]}}}}}},"/auth/google":{"get":{"tags":["Auth"],"description":"action desccription","operationId":"redirectToGoogleAuth","summary":"redirectToGoogleAuth","security":[{"BearerAuth":[]}],"parameters":[]}},"/auth/facebook":{"get":{"tags":["Auth"],"description":"action desccription","operationId":"redirectToFacebookAuth","summary":"redirectToFacebookAuth","security":[{"BearerAuth":[]}],"parameters":[]}},"/auth/login":{"post":{"tags":["Auth"],"description":"action desccription","operationId":"login","summary":"login","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"properties":{"email":{"minLength":1,"maxLength":255,"type":"string","format":"email"},"password":{"minLength":6,"maxLength":20,"type":"string"}},"type":"object","required":["email","password"]}}}}}},"/auth/google/callback":{"get":{"tags":["Auth"],"description":"action desccription","operationId":"googleCallback","summary":"googleCallback","security":[{"BearerAuth":[]}],"parameters":[]}},"/auth/accounts":{"get":{"tags":["Auth"],"description":"action desccription","operationId":"getAccountList","summary":"getAccountList","security":[{"BearerAuth":[]}],"parameters":[{"in":"query","description":"paramester description","required":false,"content":{"application/json":{"schema":{"properties":{"filter":{"properties":{"type":{"enum":["AND","OR"],"type":"string"},"filters":{"items":{"properties":{"property":{"type":"string"},"operator":{"type":"string","enum":["eq","lt","lte","gt","gte","ne","in","like","between"]},"value":{}},"type":"object","required":["property","operator"]},"type":"array"}},"type":"object","required":["type","filters"]},"sorts":{"items":{"type":"string"},"type":"array"},"limit":{"minimum":1,"type":"number"},"page":{"minimum":1,"type":"number"}},"type":"object"}}}}]}}},"components":{"schemas":{"AuthLoginRequestProps":{"properties":{"email":{"minLength":1,"maxLength":255,"type":"string","format":"email"},"password":{"minLength":6,"maxLength":20,"type":"string"}},"type":"object","required":["email","password"]},"AuthRegisterRequestProps":{"properties":{"email":{"minLength":1,"maxLength":255,"type":"string","format":"email"},"password":{"minLength":6,"maxLength":20,"type":"string"},"roleId":{"format":"uuid","type":"string"}},"type":"object","required":["email","password","roleId"]},"FilterItem":{"properties":{"property":{"type":"string"},"operator":{"type":"string","enum":["eq","lt","lte","gt","gte","ne","in","like","between"]},"value":{}},"type":"object","required":["property","operator"]},"Filter":{"properties":{"type":{"enum":["AND","OR"],"type":"string"},"filters":{"items":{"$ref":"#/components/schemas/FilterItem"},"type":"array"}},"type":"object","required":["type","filters"]},"MixedFilter":{"properties":{"type":{"enum":["AND","OR"],"type":"string"},"filters":{"items":{"$ref":"#/components/schemas/Array"},"type":"array"}},"type":"object","required":["type","filters"]},"GetQuery":{"properties":{"filter":{"$ref":"#/components/schemas/Filter"},"sorts":{"items":{"type":"string"},"type":"array"},"limit":{"minimum":1,"type":"number"},"page":{"minimum":1,"type":"number"}},"type":"object"},"GetMixedQuery":{"properties":{"filter":{},"sorts":{},"limit":{},"page":{},"search":{"type":"string"}},"type":"object","required":["search"]},"TestCreateRequestProps":{"properties":{"name":{"minLength":1,"maxLength":255,"type":"string"},"description":{"minLength":1,"maxLength":255,"type":"string"}},"type":"object","required":["name"]},"TestUpdateRequestProps":{"properties":{"name":{"minLength":1,"maxLength":255,"type":"string"},"description":{"minLength":1,"maxLength":255,"type":"string"}},"type":"object","required":["name"]},"TestDeleteRequestProps":{"properties":{"ids":{"items":{"type":"string"},"type":"array"}},"type":"object","required":["ids"]}},"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"definitions":{"AuthLoginRequestProps":{"properties":{"email":{"minLength":1,"maxLength":255,"type":"string","format":"email"},"password":{"minLength":6,"maxLength":20,"type":"string"}},"type":"object","required":["email","password"]},"AuthRegisterRequestProps":{"properties":{"email":{"minLength":1,"maxLength":255,"type":"string","format":"email"},"password":{"minLength":6,"maxLength":20,"type":"string"},"roleId":{"format":"uuid","type":"string"}},"type":"object","required":["email","password","roleId"]},"FilterItem":{"properties":{"property":{"type":"string"},"operator":{"type":"string","enum":["eq","lt","lte","gt","gte","ne","in","like","between"]},"value":{}},"type":"object","required":["property","operator"]},"Filter":{"properties":{"type":{"enum":["AND","OR"],"type":"string"},"filters":{"items":{"$ref":"#/components/schemas/FilterItem"},"type":"array"}},"type":"object","required":["type","filters"]},"MixedFilter":{"properties":{"type":{"enum":["AND","OR"],"type":"string"},"filters":{"items":{"$ref":"#/components/schemas/Array"},"type":"array"}},"type":"object","required":["type","filters"]},"GetQuery":{"properties":{"filter":{"$ref":"#/components/schemas/Filter"},"sorts":{"items":{"type":"string"},"type":"array"},"limit":{"minimum":1,"type":"number"},"page":{"minimum":1,"type":"number"}},"type":"object"},"GetMixedQuery":{"properties":{"filter":{},"sorts":{},"limit":{},"page":{},"search":{"type":"string"}},"type":"object","required":["search"]},"TestCreateRequestProps":{"properties":{"name":{"minLength":1,"maxLength":255,"type":"string"},"description":{"minLength":1,"maxLength":255,"type":"string"}},"type":"object","required":["name"]},"TestUpdateRequestProps":{"properties":{"name":{"minLength":1,"maxLength":255,"type":"string"},"description":{"minLength":1,"maxLength":255,"type":"string"}},"type":"object","required":["name"]},"TestDeleteRequestProps":{"properties":{"ids":{"items":{"type":"string"},"type":"array"}},"type":"object","required":["ids"]}}}